[SCRUM-245] Chore : 파일명 변경
